requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cssselect==1.1.0
DataProperty==0.48.2
decorator==4.4.2
fake-useragent==0.1.11
future==0.18.2
idna==2.9
lxml==4.5.0
mbstrdecoder==1.0.0
msgfy==0.1.0
parse==1.15.0
pathvalidate==2.2.2
peewee==3.13.3
pycparser==2.20
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
pytablewriter==0.51.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.6.1
pytz==2019.3
requests==2.23.0
requests-html==0.10.0
six==1.14.0
soupsieve==2.0
tabledata==1.1.2
tornado==6.0.4
tqdm==4.45.0
typepy==1.1.0
urllib3==1.25.9
w3lib==1.21.0
websockets==8.1
